PROJECT NOTION  PRES PARC/presparc-notion/requirements.txt
notion-client>=2.2.1
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0
rich>=13.0.0